+++ pkg/node/node.go
@@ -183,8 +183,8 @@ const (
 	minPaymentThreshold           = 2 * refreshRate           // minimal accepted payment threshold of full nodes
 	maxPaymentThreshold           = 24 * refreshRate          // maximal accepted payment threshold of full nodes
 	mainnetNetworkID              = uint64(1)                 //
-	ReserveCapacity               = 16384                     // 2^14 chunks
-	reserveWakeUpDuration         = 15 * time.Minute          // time to wait before waking up reserveWorker
+	ReserveCapacity               = 65536                     // 2^14 chunks
+	reserveWakeUpDuration         = 5 * time.Minute           // time to wait before waking up reserveWorker
 	reserveTreshold               = ReserveCapacity * 5 / 10
 	reserveMinimumRadius          = 0
 )
